+++ goodtables/cli.py
@@ -10,27 +10,30 @@ import json as json_module
 from pprint import pformat
 from datapackage import Package
 from click_default_group import DefaultGroup
+from .helpers import init_datapackage
 from . import config
 click.disable_unicode_literals_warning = True
 
 
 # Module API
 
-@click.group(cls=DefaultGroup, default='validate', default_if_no_args=True)
+@click.group(cls=DefaultGroup, default='validate', default_if_no_args=True, help='')
 @click.version_option(config.VERSION, message='%(version)s')
 def cli():
-    """Tabular files validator.
+    """Command-line interface
 
-    There are two categories of validation checks available:
+    ```
+    Usage: cli.py [OPTIONS] COMMAND [ARGS]...
 
-    * Structural checks: ensure there are no empty rows, no blank headers, etc.
+    Options:
+      --version  Show the version and exit.
+      --help     Show this message and exit.
 
-    * Content checks: ensure the values have the correct types (e.g. string),
-      their format is valid (e.g. e-mail), and they respect some constraint
-      (e.g. age is greater than 18).
+    Commands:
+      validate*  Validate tabular files (default).
+      init       Init data package from list of files.
+    ```
 
-    \b
-    Full documentation at: <https://github.com/frictionlessdata/goodtables-py/>
     """
     pass
 
@@ -171,24 +174,3 @@ def _print_report(report, output=None, json=False):
             }
             message = template.format(**substitutions)
             secho(message)
-
-
-# Internal
-
-def init_datapackage(resource_paths):
-    # "Create tabular data package with resources.
-    dp = Package({
-        'name': 'change-me',
-        'schema': 'tabular-data-package',
-    })
-
-    for path in resource_paths:
-        dp.infer(path)
-
-    return dp
-
-
-# Main program
-
-if __name__ == '__main__':
-    cli()
